+++ backend/src/presentation/controllers/locality/load-ufs/load-ufs-controller.spec.ts
@@ -1,7 +1,7 @@
-import { mockUfsModel } from './../../domain/test/mock-locality';
-import { LoadUfsSpy } from './../test/mock-locality';
-import { ServerError } from './../errors/server-error';
-import { serverError, forbidden, ok } from './../helpers/http/http-helper';
+import { mockUfsModel } from '../../../../domain/test/mock-locality';
+import { LoadUfsSpy } from '../../../test/mock-locality';
+import { ServerError } from '../../../errors/server-error';
+import { serverError, forbidden, ok } from '../../../helpers/http/http-helper';
 import { LoadUfsController } from './load-ufs-controller';
 
 type SutTypes = {
